+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 
 - permit to force gate / duration to operator (restapi add booking)
   [mamico]
+- Changes required to migrate the old bookings.
+  [folix-01]
+
 
 2.0.0.dev4 (2023-08-11)
 -----------------------